requirements.txt
autopep8==1.4.4
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.1
django-bootstrap4==1.0.1
django-friendship==1.8.1
django-tinymce==2.8.0
psycopg2==2.8.4
pycodestyle==2.5.0
soupsieve==1.9.5
